cabwera_ecommerce/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2
django-crispy-forms==2.4
django-registration-redux==2.13
django-tinymce==4.1.0
django-versatileimagefield==3.1
pillow==11.2.1
psycopg==3.2.7
python-decouple==3.8
python-magic==0.4.27
setuptools==80.1.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
